Move config.js
