+++ app.js
@@ -4,7 +4,7 @@ const logger = require('morgan')
 const cookieParser = require('cookie-parser')
 const bodyParser = require('body-parser')
 const mongoose = require('mongoose')
-const config = require('./config')
+const config = require('./config/main')
 
 const index = require('./routes/index')
